pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.1.0"
description = ""
authors = ["bot <kotohawk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
django-extensions = "^3.1.5"
social-auth-app-django = "^5.0.0"
pytest = "^7.1.1"
pytest-django = "^4.5.2"
fontawesomefree = "^6.1.1"
Pillow = "^9.0.1"
freezegun = "^1.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
